# LUT-25878 : Remove the EntryTypeServiceManager Cache