This cache is used for caching spring managed singleton beans. The spring application context is already a cache for those. Testing performance :
Before:
Querying 30 entries 250000 times took 152ms
Querying 30 entries 250000 times took 43ms
Querying 30 entries 250000 times took 35ms
Querying 30 entries 250000 times took 35ms
Querying 30 entries 250000 times took 35ms
Querying 30 entries 250000 times took 35ms
After:
Querying 30 entries 250000 times took 46ms
Querying 30 entries 250000 times took 24ms
Querying 30 entries 250000 times took 11ms
Querying 30 entries 250000 times took 11ms
Querying 30 entries 250000 times took 10ms
Querying 30 entries 250000 times took 11ms

So removing the cache improves performance